Move all other components to the components folder Add a new line  in thetoggle file
